Threading fixed in mock devices
