+++ example_scripts/finite_multi_fifo_mode.py
@@ -70,7 +70,7 @@ if __name__ == "__main__":
     from matplotlib.pyplot import plot, show, figure, title
 
     measurements = finite_multi_fifo_example(
-        mock_mode=True, num_measurements=2, trigger_source=TriggerSource.SPC_TMASK_EXT0, device_number=0
+        mock_mode=True, num_measurements=5, trigger_source=TriggerSource.SPC_TMASK_EXT0, device_number=0
     )
 
     # Plot waveforms
@@ -78,7 +78,7 @@ if __name__ == "__main__":
         figure()
         title(f"Measurement {n}")
         for waveform in measurement:
-            plot(waveform)
+            plot(waveform.samples)
 
     print(f"Completed {len(measurements)} measurements each containing {len(measurements[0])} waveforms.")
     print(f"Waveforms had the following shape: {measurements[0][0].samples.shape}")
