+++ setup.cfg
@@ -31,6 +31,7 @@ zip_safe = False
 packages =
   spectrumdevice
   spectrumdevice.devices
+  spectrumdevice.devices.mocks
   spectrumdevice.settings
   spectrumdevice.spectrum_wrapper
   spectrum_gmbh
