+++ spectrumdevice/__init__.py
@@ -27,8 +27,7 @@ from .devices.spectrum_device import SpectrumDevice
 from .devices.spectrum_card import SpectrumCard
 from .devices.spectrum_channel import SpectrumChannel
 from .devices.spectrum_star_hub import SpectrumStarHub
-from spectrumdevice.devices.mocks.mock_devices import MockSpectrumCard
-from spectrumdevice.devices.mocks.mock_devices import MockSpectrumStarHub
+from spectrumdevice.devices.mocks import MockSpectrumCard, MockSpectrumStarHub
 
 __all__ = [
     "SpectrumDevice",
