+++ tests/configuration.py
@@ -12,8 +12,8 @@ class SpectrumTestMode(Enum):
 
 # Set to TestMode.MOCK_HARDWARE for software-only testing, even if Spectrum drivers are found on the system
 # Set to TestMode.REAL_HARDWARE to run tests on a real hardware device as configured below.
-SINGLE_CARD_TEST_MODE = SpectrumTestMode.REAL_HARDWARE
-STAR_HUB_TEST_MODE = SpectrumTestMode.REAL_HARDWARE
+SINGLE_CARD_TEST_MODE = SpectrumTestMode.MOCK_HARDWARE
+STAR_HUB_TEST_MODE = SpectrumTestMode.MOCK_HARDWARE
 
 # Set IP address of real spectrum device (for use if TestMode.REAL_HARDWARE is set above). Set to None to run tests on
 # a local (PCIe) card.
